Refactored financial reported pages
