+++ src/finance_modules/SpendingMonthlyBarChart.jsx
